refactor: tiger style compliance for aspen-node.rs
Replace usize with u32 across 9 files for portability. Refactor metrics(),
main(), and health() functions from 637 lines to 105 lines by extracting 18
focused helper functions. Remove production .expect() by introducing
DEFAULT_HTTP_ADDR constant. All functions now <70 lines.

Tests: 211/211 passing

+++ src/cluster/config.rs
@@ -94,10 +94,10 @@ pub struct ClusterBootstrapConfig {
     /// Default: 1000 messages
     /// Maximum: 10000 messages
     #[serde(default = "default_raft_mailbox_capacity")]
-    pub raft_mailbox_capacity: usize,
+    pub raft_mailbox_capacity: u32,
 }
 
-fn default_raft_mailbox_capacity() -> usize {
+fn default_raft_mailbox_capacity() -> u32 {
     1000
 }
